Remove jupyter for now
There is a dependency conflict between jupyter and vizier that pip can not resolve it appears. So jupyter has to be installed after vizier to have the correct package installed. Let's remove jupyter for now then and users can pip install jupyter themselves if them want to use the notebooks. We can create a page on the documentation for notebooks and make sure this step is reflected. Sound good @srivatsankrishnan?